// src/types/circuit.ts
export interface Circuit {
  id: number;
  name: string;
  slug: string;
  reference: string;
  lat: number;
  lng: number;
  country: string;
  country_iso: string;
  city: string;
  address: string;
  phone: string;
  website: string;
  photo_url: string;
  photo_thumb_url: string;
  sws_url: string;
  currency: string;
}